Fixed a line of code for bills
