+++ client/src/components/Navbar.js
@@ -24,11 +24,11 @@ function Navbar() {
               Bills
             </Link>
           </li>
-          <li>
+          {/* <li>
             <Link to="/Billss" className="hover:text-gray-300">
               Billss
             </Link>
-          </li>
+          </li> */}
           <li>
             <Link to="/donation" className="hover:text-gray-300">
               Donation
